Move Semester table access from school_mappers to views
Currently we manually define the available semesters for each school, but
this might lead to a scenario where we get a request for a semester that
does not exist in the DB yet. The responsibility to ensure that the db
is up to date with school_mappers is moved to views.py to avoid making
database queries when importing from school_mappers (or any module
that imports from school_mappers).