Custom Widget (CSS/JS/PY) working, but not HTML
A custom widget was built for the timetable. The data is being fetched
properly on both the backend and frontend, but it is not being rendered
in the div on the HTML. This should hopefully be fixed soon, but a
commit is being made for now, since some progress was made on custom
widgets.