Use new CBVs for profile management
Namely, views for saving settings and for the user profile page.
Since these refer to the same resource (the user profile), they now have
the same endpoint - the profile page is returned on a GET request and saving
settings is accepted through a PATCH request, in accordance with RESTful
API design.

In order to do this, code for the old endpoint (timetable.views.profile) had
to be moved to the student app, alongside the PATCH logic, which then required
a rearranging of certain functions in analytics, timetable, and student in order
to avoid circular imports.