Use POST to update saved timetables
The original code for PUT missed an edge case - if a timetable is renamed
to an existing name, there should be a conflict. Without passing in the old
name from the frontend (which requires new functionality to be added), the
PUT handler is essentially the same as the POST handler. Since POST is also
an acceptable method for updating data, the two have been merged together.