Provide user object for middleware mixins when testing with request factory
Using rest framework's APIView means that the authentication pipeline is run
inside of the view's dispatch function (before calling the view handler), which
is different from normal django CBVs.
This means that middleware mixins like LoginRequiredMixin will not have a
.user attribute set on requests when using a request factory when they override
the dispatch method. To get around this, we provide the user when testing with
request factories.

The reason this approach is used instead of the RESTful way of responding
with 403s/401s (which rest-framework supports), is that this requires some level
of boilerplate on both the frontend (error handle response based on status code)
as well as the backend (modify the authentication/permissions of the view).
Since our API is currently only used by the frontend, we do it the easy way
rather than the RESTful way until this is an issue.