Add handler for feature flows with old semesters
When a feature flow provides a semester that is not up to date (for example
sharing a course or a timetable from 2 years ago), or more precisely if the
provided semester is not in the list returned by `get_current_semesters()`,
then this semester is appended to `allSemesters` state provided to the frontend.

This approach was chosen since we want users to be able to load data
from previous years, and this approach allows that with the minimal amount
of changes and since solutions which involve separating redux semester state with the
semester of the feature flow object lead to issues with consistency. For example,
if a shared timetable is loaded and the semester is not changed to reflect that shared
timetable, there is no way for the user to know which semester the shared
timetable is from.