Move "helper" functions to endpoints
Some functions which return a url are in helpers dir instead of in endopints.jsx
For the second time, not having all url related functions in the same file has
led to bugs since someone making changes which affect all or some urls has not
checked the helper files for url functions. This change addresses that.