Change cached semester format
Previously we stored the index into state.semester.all when caching the
semester. However this makes the index invalid when state.semester.all changes
(for e.g. when we add semesters for the next school year). Using an index
does not give us enough information to keep cached timetables after the state
changes. Therefore we cache both the semester name and year, and look up
the correct index in the state.
This means that users will lose their cached timetable, which is necessary
to apply this fix across all schools.