Temporarily remove course modal body snapshot test
Course modal body is a component with complex props that can't be feasibly
mocked by hand and therefore requires a fixture. Currently there is no
easy way to produce these fixtures of props besides manually loading the app
in the browser and copy and pasting props. This defeats the purpose of the
automated test if it is necessary to load the page manually every time the
props are changed.
Therefore the test is removed until there is an easier way to update fixtures or
the normalizing refactor is finished (where many state shape changes will be made).