Celery tasks synced with school_mappers
Scheduled tasks will now run in accordance to data in school_mappers. There
are scheduled tasks that can be set to run all active semesters as well
as tasks that can be set to run only the most recent semesters. This means, as
well, that parallelism is now seen not only per school but per semester. There
are edges that handle schools that can only have a single login (GW) as well
as logic to handle schools that have two current/recent semesters (QUEENS).