Revert ptt response shape
Instead of responding with 204 no content, respond with the changed timetable
as well as the entire list of personal timetables as before.
Responding with 204 no content would need the frontend to implement a way to update
userInfo.timetables (add/edit, and delete), and then reload the active timetable.
In order to do this, the frontend would need to assume that names are unique
which might not be true if the user has more than one instance of semesterly
open, or the backend would need to send back database ids.
The fact that one timetable is found in userInfo, savingTimetable, and timetables
would benefit from normalizing timetables. However, this would either be extra work
for timetables that will never be found in multiple parts of state (generated
or shared timetables), or we would need to have specific handling for personal
timetables so that ptt and shared tt IDs don't get mixed up.
All in all, responding with 204 no content or even just the saved/edited/
deleted personal timetable would require a lot more work which should be tackled
separately later.